requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
cramjam==2.9.1
exceptiongroup==1.2.2
fastparquet==2024.11.0
filelock==3.16.1
fsspec==2024.12.0
huggingface-hub==0.27.1
idna==3.10
# Editable Git install with no remote (image-text-validator==0.1.0)
-e /Users/vassilyn/PycharmProjects/image_description_validator
iniconfig==2.0.0
jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.24.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.48.0
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
